requirements.txt
fastapi
uvicorn[standard]
pymysql
alembic
sqlalchemy
pydantic[email]
bcrypt
python-dotenv
pydantic-settings
pyright
aiohttp
orjson
python-jose[cryptography]
passlib[bcrypt]
